Fix: realizando os ajustes finais naapi

+++ src/service/postagem.service.js
@@ -1,12 +1,14 @@
 import { PostagemRepository } from "../repository/postagem.repository.js";
+import { UsuarioRepository } from "../repository/usuario.repository.js";
 
 export class PostagemService {
 
     postagemRepository = new PostagemRepository();
+	usuarioRepository = new UsuarioRepository();
 
-    buscar = async () => {
+	buscar = async (filtros) => {
         try {
-            const { resultado: listaPostagens } = await this.postagemRepository.buscarPostagens({});
+			const { resultado: listaPostagens } = await this.postagemRepository.buscarPostagens(filtros);
 
             return {
                 status: 200,
@@ -71,11 +73,43 @@ export class PostagemService {
 
     editar = async (postagem) => {
         try {
+			const { possuiResultado: postagemEncontrada } =
+				await this.postagemRepository.buscarPostagens({
+					id: postagem.id,
+				});
+
+			if (!postagemEncontrada) {
+				return {
+					status: 400,
+					resposta: {
+						mensagem: `Postagem ${postagem.id} não foi encontrada`,
+					},
+				};
+			}
+
+			if (postagem.usuarioId) {
+				const { possuiResultado: usuarioEncontrado } =
+					await this.usuarioRepository.buscarUsuarios({
+						id: postagem.usuarioId,
+					});
+				
+				if(!usuarioEncontrado) {
+					return {
+						status: 400,
+						resposta: {
+							mensagem: `Usuário ${postagem.usuarioId} não foi encontrado`
+						}
+					}
+				}
+			}
+
             const postagemEditada = await this.postagemRepository.editarPostagem(postagem);
 
             return {
                 status: postagemEditada ? 200 : 500,
-                resposta: postagemEditada ? 'Postagem editada com sucesso' : 'Erro ao editar a postagem',
+                resposta: {
+					mensagem: postagemEditada ? 'Postagem editada com sucesso' : 'Erro ao editar a postagem'
+				},
             };
         } catch (erro) {
             console.log('[POSTAGEM SERVICE] Erro ao editar uma postagem', erro);
@@ -91,11 +125,26 @@ export class PostagemService {
 
     remover = async (id) => {
         try {
+			const { possuiResultado: postagemEncontrada } = 
+				await this.postagemRepository.buscarPostagens({
+					id: id,
+				});
+			
+			if(!postagemEncontrada) {
+				return {
+					status: 400,
+					resposta: {
+						mensagem: `Postagem ${id} não encontrada`,
+					},
+				};
+			}
             const postagemRemovida = await this.postagemRepository.removerPostagem(id);
 
             return {
                 status: postagemRemovida ? 200 : 500,
-                resposta: postagemRemovida ? 'Postagem removida com sucesso' : 'Erro ao remover a postagem',
+                resposta: {
+					mensagem: postagemRemovida ? 'Postagem removida com sucesso' : 'Erro ao remover a postagem',
+				},
             };
         } catch (erro) {
             console.log('[POSTAGEM SERVICE] Erro ao remover uma postagem', erro);
@@ -107,29 +156,5 @@ export class PostagemService {
                 }
             };
         }
-    };
-
-    buscarPorFiltros = async (filtros) => {
-        try {
-            let { resultado: listaPostagens } = await this.postagemRepository.buscarPostagens(filtros);
-
-            if(filtros.id) {
-                listaPostagens = listaPostagens ? [listaPostagens] : [];
-            }
-
-            return {    
-                status: 200,
-                resposta: listaPostagens,
-            };
-        } catch (erro) {
-            console.log('[POSTAGEM SERVICE] Erro ao buscar postagem por filtros', erro);
-
-            return {
-                status: 500,
-                resposta: {
-                    mensagem: "Erro durante a busca das postagens",
-                }
-            };
-        }
-    };
+	};
 }
